jimt/dance-ai-modal-refactoring - pulled all functions out of DanceAiModal
