+++ apps/src/dance/ai/utils/getImageUrl.ts
@@ -1,0 +1,3 @@
+export const getImageUrl = (id: string) => {
+  return `/blockly/media/dance/ai/emoji/${id}.svg`;
+};
