+++ apps/src/dance/ai/utils/index.js
@@ -1,5 +1,9 @@
 export * from './generateBlocks';
 export * from './generateBlocksFromResult';
 export * from './generatePreviewCode';
+export * from './getImageUrl';
 export * from './getLabelMap';
+export * from './getRangeArray';
+export * from './linearInterpolation';
+export * from './useInterval';
 export * from './validateAndSetFieldValue';
