+++ apps/src/dance/ai/utils/linearInterpolation.ts
@@ -1,0 +1,4 @@
+// Linear interpolation.
+export const lerp = (step: number, steps: number, min: number, max: number) => {
+  return (step / steps) * (max - min) + min;
+};
