+++ apps/src/dance/ai/utils/useInterval.ts
@@ -1,0 +1,24 @@
+import {useEffect, useRef} from 'react';
+
+// Adapted from https://overreacted.io/making-setinterval-declarative-with-react-hooks/
+export function useInterval(callback: () => void, delay: number | undefined) {
+  const savedCallback = useRef<() => void>();
+
+  // Remember the latest callback.
+  useEffect(() => {
+    savedCallback.current = callback;
+  }, [callback]);
+
+  // Set up the interval.
+  useEffect(() => {
+    function tick() {
+      if (savedCallback.current) {
+        savedCallback.current();
+      }
+    }
+    if (delay !== undefined) {
+      const id = setInterval(tick, delay);
+      return () => clearInterval(id);
+    }
+  }, [delay]);
+}
